pyproject.toml
[project]
name = "pangea-anthropic"
version = "0.0.0"
description = "A wrapper around the Anthropic Python library"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Pangea", email = "support@pangea.cloud" }
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.64.0",
    "anyio>=4.10.0",
    "httpx>=0.28.1",
    "pangea-sdk>=6.5.0",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "mypy==1.17.1",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "ruff==0.12.11",
]

[build-system]
requires = ["uv_build==0.8.8"]
build-backend = "uv_build"

[tool.mypy]
python_version = "3.10"
pretty = true
show_error_codes = true
implicit_reexport = false
cache_fine_grained = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",   # isort
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TC",  # flake8-type-checking
    "UP",  # pyupgrade
]
ignore = [
    "TC001",
    "TC002",
    "TC003",

    # Reduce the delta from anthropic-sdk-python.
    "E501",
    "UP006",
    "UP007",
    "UP035",
]
